# Conflicts:
#	src/main/java/com/example/catsanddogstelegram/repository/CatRepository.java
#	src/main/java/com/example/catsanddogstelegram/repository/DogRepository.java

+++ src/main/java/com/example/catsanddogstelegram/repository/DogRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DogRepository extends JpaRepository<Dog, Integer> {
+public interface DogRepository extends JpaRepository<Dog, Long> {
 }
